requirements2.txt
asgiref==3.5.2
attrs==22.1.0
Brotli==1.0.9
cached-property==1.5.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
cssselect2==0.7.0
defusedxml==0.7.1
Django==3.2
django-cors-headers==3.13.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
fonttools==4.38.0
fpdf==1.7.2
gunicorn==20.1.0
html5lib==1.1
idna==3.3
isodate==0.6.1
itypes==1.2.0
Jinja2==3.1.2
lxml==4.9.1
mailjet-rest==1.3.4
MarkupSafe==2.1.1
mysqlclient==2.1.1
oauthlib==3.2.0
Pillow==9.3.0
platformdirs==2.5.2
pycparser==2.21
pycryptodome==3.15.0
pydyf==0.5.0
PyJWT==2.4.0
pyphen==0.13.0
python3-openid==3.2.0
pytz==2022.1
qrcode==7.3.1
requests==2.28.1
requests-file==1.5.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
suds==1.1.2
tinycss2==1.2.1
uritemplate==4.1.1
urllib3==1.26.9
weasyprint==57.0
webencodings==0.5.1
xmltodict==0.13.0
zeep==4.1.0
zopfli==0.2.1
